re-usable Auth structs and functions moved to `lib`.
